client-plot/Generate/requirements.txt
numpy==1.26.4
matplotlib
supabase
pydantic
fastapi[standard]
pandas
python-dotenv
